Use OpenJDK 6 as prereq
The libmesos library includes an RPATH entry, which is needed to find
libjvm.so at runtime.  This RPATH is hard-coded to the OpenJDK version
that was present when the package was compiled.  So even though the
Debian package claims that it works with either OpenJDK 6 or OpenJDK 7,
the fact that Mesosphere compiled with OpenJDK 6 means that we have to
have that specific version present at runtime.